+++ src/app/inventory/[topLevelContainerId]/[itemType]/[itemId]/layout.tsx
@@ -3,21 +3,9 @@ import { ItemLayoutContent } from "./layoutContent";
 import React from "react";
 import { getShipmentData } from "@/utils/server/shipment";
 import { TreeData } from "@/components/visualisation/treeView";
-import { components } from "@/types/schema";
 import { InventoryItemLayoutProps } from "@/types/generic";
-import { VStack, Heading, Link, Text } from "@chakra-ui/react";
+import { VStack, Heading, Text, Link } from "@chakra-ui/react";
 import NextLink from "next/link";
-
-const getUnassignedItems = async () => {
-  const res = await serverFetch("/internal-containers/unassigned", {
-    cache: "force-cache",
-    next: { tags: ["shipment"] },
-  });
-
-  return res && res.status === 200
-    ? ((await res.json()) as components["schemas"]["Paged_GenericItem_"])
-    : null;
-};
 
 const ItemLayout = async (props: InventoryItemLayoutProps) => {
   const params = await props.params;
@@ -30,9 +18,7 @@ const ItemLayout = async (props: InventoryItemLayoutProps) => {
     "topLevelContainer",
   )) as TreeData;
 
-  const unassignedItems = await getUnassignedItems();
-
-  if (shipmentData === null || unassignedItems === null) {
+  if (shipmentData === null) {
     return (
       <VStack w='100%' mt='3em'>
         <Heading variant='notFound'>Inventory Item Unavailable</Heading>
@@ -43,11 +29,7 @@ const ItemLayout = async (props: InventoryItemLayoutProps) => {
   }
 
   return (
-    <ItemLayoutContent
-      params={params}
-      shipmentData={shipmentData}
-      unassignedItems={unassignedItems}
-    >
+    <ItemLayoutContent params={params} shipmentData={shipmentData}>
       {children}
     </ItemLayoutContent>
   );