+++ src/mappings/forms/dewar.ts
@@ -6,17 +6,20 @@ export const dewarForm = [
   {
     id: "code",
     label: "Dewar Code",
-    type: "editableDropdown",
-    validation: {
-      maxLength: { value: 15, message: "Serial number have less than 15 characters" },
-    },
-    hint:
-      "Select your dewar's code, or generate a new one if your dewar has no associated code." +
-      "\nIf generating a new code, serial numbers are required for international shipments." +
-      "\nIf your dewar has no serial code, set the value to N/A.",
+    type: "dropdown",
+    hint: "Select your dewar's code, or generate a new one if your dewar has no associated code.",
     values: {
       $ref: { parent: "#/dewars", map: { value: "facilityCode", label: "facilityCode" } },
-      base: [{ label: "Generate New Code (Specify Serial Number)", value: "other" }],
+      base: [{ label: "Generate New Code", value: null }],
+    },
+  },
+  {
+    id: "manufacturerSerialNumber",
+    label: "Manufacturer Serial Code",
+    type: "text",
+    hint: "If your dewar has no serial code, leave this blank. The serial code must match the code on the dewar.",
+    validation: {
+      maxLength: { value: 15, message: "Serial number have less than 15 characters" },
     },
   },
 ] as DynamicFormEntry[];