+++ src/utils/generic.tsx
@@ -103,6 +103,7 @@ export const pascalToSpace = (input: string) =>
   input.replace(/([A-Z])/g, " $1").replace(/^./, (str) => str.toUpperCase());
 
 export const nameValidation = {
+  maxLength: 80,
   pattern: {
     value: /^[a-zA-Z0-9_]*$/,
     message: "Name must only contain alphanumeric characters and underscores",